day1128/requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.3
django-tinymce==4.1.0
djangorestframework==3.15.2
Markdown==3.7
mysqlclient==2.2.4
pillow==11.0.0
Pygments==2.18.0
sqlparse==0.5.1

# Python 3.11.2